Restructure app into functional subfolders for easier navigability
