+++ client/src/components/media/retrieve/ImageRetrieve.js
