+++ client/src/routes.js
@@ -1,15 +1,14 @@
 //File contains the route setup to be exported to be used by App.js
 import React from 'react'
 import { Router, Route, hashHistory, IndexRoute } from 'react-router'
-import HomeContainer from './containers/HomeContainer'
-import MyMemlysContainer from './containers/MyMemlysContainer'
-import LikedMemlysContainer from './containers/LikedMemlysContainer'
-import ProfileContainer from './containers/ProfileContainer'
-import EditProfileContainer from './containers/EditProfileContainer';
-import ImageUploadContainer from './containers/ImageUploadContainer'
-import GoogleMapContainer from './containers/GoogleMapContainer';
+import HomeContainer from './components/home/HomeContainer'
+import MyMemlysContainer from './components/user/myMemlys/MyMemlysContainer'
+import LikedMemlysContainer from './components/user/likedMemlys/LikedMemlysContainer'
+import ProfileContainer from './components/user/profile/ProfileContainer'
+import EditProfileContainer from './components/user/editProfile/EditProfileContainer';
+import ImageUploadContainer from './components/media/upload/ImageUploadContainer'
+import GoogleMapContainer from './components/map/map/GoogleMapContainer';
 import axios from 'axios'
-
 
 var routes = (
   <Router history={hashHistory}>
@@ -26,9 +25,7 @@ var routes = (
   </Router>
 )
 
-
-
-// you can use an onEnter hook before enteing routes to check if user is authorized. (i.e. can check the state logged in or not)
+// you can use an onEnter hook before entering routes to check if user is authorized. (i.e. can check the state logged in or not)
 // this is a good way to do client side route protection. Without this, although they may not be able to see any of their data, anyone can still enter /user/profile route
 function checkAuth(nextState, replace){
   //the fact that we have been rerouted by server to here means that we are authorized to be here
@@ -39,4 +36,3 @@ function checkAuth(nextState, replace){
 }
 
 export default routes
-
